requirements.txt
Django==4.2.25
psycopg2-binary==2.9.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
requests>=2.31.0
python-dotenv>=1.0.0
